some ranaming and separate state initializer
