+++ admin-dev/themes/new-theme/.webpack/common.js
@@ -100,7 +100,7 @@ module.exports = {
     translations: './js/app/pages/translations',
     webservice: './js/pages/webservice',
     theme: ['./scss/theme.scss'],
-    product: './js/pages/product',
+    product_edit: './js/pages/product/edit',
   },
   output: {
     path: path.resolve(__dirname, '../public'),
