+++ admin-dev/themes/new-theme/js/pages/product/product-map.js
