+++ src/PrestaShopBundle/Controller/Admin/ProductController.php
@@ -125,7 +125,7 @@ class ProductController extends FrameworkBundleAdminController
         $sortOrder = 'desc'
     ) {
         //@todo;
-        return $this->render('@PrestaShop/Admin/Sell/Catalog/Product/index.html.twig');
+        return $this->render('@PrestaShop/Admin/Sell/Catalog/Product/edit.html.twig');
 
         if (!$this->isGranted([PageVoter::READ, PageVoter::UPDATE, PageVoter::CREATE], self::PRODUCT_OBJECT)) {
             return $this->redirect('admin_dashboard');
